Query entity search result life spans when concept is Person.
An additional SPARQL query is transmitted during a wikidata entity search that requests birth and death dates of all
search results of the correct type, if the concept used for classification is Person.

If we search for an entity classified as any other concept, no such query is sent. If the entity we are looking for
ought to be a Person and a query is sent, it will exclude any search results that are already known to be of the wrong
type from the previous step.